stress: use nanosleep() instead of usleep()
The docs mention that usleep() is obsolete in newer POSIX specs.
And nanosleep() should be used.

It's not clear whether there is any libc out there that actually removed
usleep().
In my case, I just maintain stress (and stress-ng) in OpenWrt and this
patch is in our tree for a while:
  https://github.com/openwrt/packages/blob/3218dd24577b051fddb1635a5a856fe8426ef8f8/utils/stress/patches/010-usleep.patch

Refs:
  https://man7.org/linux/man-pages/man3/usleep.3.html

   4.3BSD, POSIX.1-2001.  POSIX.1-2001 declares this function
   obsolete; use nanosleep(2) instead.  POSIX.1-2008 removes the
   specification of usleep().

Signed-off-by: Alexandru Ardelean <alex@shruggie.ro>